% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnePrimerRemove.R
\name{OnePrimerRemove}
\alias{OnePrimerRemove}
\title{Clean biological secuences of one primer}
\usage{
OnePrimerRemove(SEQ, PrimerOPR)
}
\arguments{
\item{SEQ}{dnastring or file with fasta format containing biological sequences that are to be cleaned.}

\item{PrimerOPR}{dnastring or file with fasta format containing the reverse primer/vector sequences to be removed.}
}
\value{
clean biological sequences
}
\description{
Cleans biological sequences of one primer.
}
\examples{
SEQ4 = DNAString("AATCG")
SEQ5 = DNAString("CG")
OnePrimerRemove (SEQ4,SEQ5)
}
\author{
Florencia I Pozzi, Gisela Y. Green, Ivana G. Barbona, Gustavo R. Rodriguez, Silvina A. Felitti
}
